// src/api/generated/models/AdditionalExpense.ts
/* tslint:disable */
/* eslint-disable */
/**
 * Blackwell Time API
 * No description provided (generated by Openapi Generator https://github.com/openapitools/openapi-generator)
 *
 * The version of the OpenAPI document: 1.0.0
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */

import { mapValues } from '../runtime';
/**
 * 
 * @export
 * @interface AdditionalExpense
 */
export interface AdditionalExpense {
    /**
     * 
     * @type {string}
     * @memberof AdditionalExpense
     */
    expenseName?: string;
    /**
     * 
     * @type {number}
     * @memberof AdditionalExpense
     */
    amount?: number;
}

/**
 * Check if a given object implements the AdditionalExpense interface.
 */
export function instanceOfAdditionalExpense(value: object): value is AdditionalExpense {
    return true;
}

export function AdditionalExpenseFromJSON(json: any): AdditionalExpense {
    return AdditionalExpenseFromJSONTyped(json, false);
}

export function AdditionalExpenseFromJSONTyped(json: any, ignoreDiscriminator: boolean): AdditionalExpense {
    if (json == null) {
        return json;
    }
    return {
        
        'expenseName': json['expenseName'] == null ? undefined : json['expenseName'],
        'amount': json['amount'] == null ? undefined : json['amount'],
    };
}

export function AdditionalExpenseToJSON(json: any): AdditionalExpense {
    return AdditionalExpenseToJSONTyped(json, false);
}

export function AdditionalExpenseToJSONTyped(value?: AdditionalExpense | null, ignoreDiscriminator: boolean = false): any {
    if (value == null) {
        return value;
    }

    return {
        
        'expenseName': value['expenseName'],
        'amount': value['amount'],
    };
}